+++ src/CollectionInputFilter.php
@@ -139,8 +139,28 @@ class CollectionInputFilter extends InputFilter
      */
     public function setData($data)
     {
-        $this->data = $data;
+        if (! (is_array($data) || $data instanceof Traversable)) {
+            throw new Exception\InvalidArgumentException(sprintf(
+                '%s expects an array or Traversable collection; invalid collection of type %s provided',
+                __METHOD__,
+                is_object($data) ? get_class($data) : gettype($data)
+            ));
+        }
 
+        foreach ($data as $item) {
+            if (is_array($item) || $item instanceof Traversable) {
+                continue;
+            }
+
+            throw new Exception\InvalidArgumentException(sprintf(
+                '%s expects each item in a collection to be an array or Traversable; '
+                . 'invalid item in collection of type %s detected',
+                __METHOD__,
+                is_object($item) ? get_class($item) : gettype($item)
+            ));
+        }
+
+        $this->data = $data;
         return $this;
     }
 
@@ -158,13 +178,11 @@ class CollectionInputFilter extends InputFilter
             }
         }
 
-        if (is_scalar($this->data)
-            || count($this->data) < $this->getCount()
-        ) {
+        if (count($this->data) < $this->getCount()) {
             $valid = false;
         }
 
-        if (empty($this->data) || is_scalar($this->data)) {
+        if (! $this->data) {
             $this->clearValues();
             $this->clearRawValues();
 
@@ -172,9 +190,6 @@ class CollectionInputFilter extends InputFilter
         }
 
         foreach ($this->data as $key => $data) {
-            if (!is_array($data)) {
-                $data = [];
-            }
             $inputFilter->setData($data);
 
             if (null !== $this->validationGroup) {
@@ -258,7 +273,7 @@ class CollectionInputFilter extends InputFilter
      */
     public function getUnknown()
     {
-        if (!is_array($this->data) || !$this->data) {
+        if (!$this->data) {
             throw new Exception\RuntimeException(sprintf(
                 '%s: no data present!',
                 __METHOD__
@@ -269,9 +284,6 @@ class CollectionInputFilter extends InputFilter
 
         $unknownInputs = [];
         foreach ($this->data as $key => $data) {
-            if (!is_array($data)) {
-                $data = [];
-            }
             $inputFilter->setData($data);
 
             if ($unknown = $inputFilter->getUnknown()) {